feat [DEV-27]: Add a new media element mock and refactor media service mock
