+++ libs/angular/ui/src/modules/media/mocks/index.ts
@@ -1,1 +1,2 @@
-export { MediaServiceMock } from './media';
+export { MediaElementMock } from './media-element.mock';
+export { MediaServiceMock } from './media-service.mock';
