--- libs/angular/ui/src/modules/media/mocks/media/index.ts
@@ -1,2 +1,0 @@
-export { MediaServiceMock } from './media-service.mock';
-export { provideMediaServiceMock } from './media-service-mock.provider';
